Finished with initial rework of auth-page ui
